requirements.txt
psycopg2-binary==2.9.9
PyPDF2==3.0.1
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3